+++ changelog.md
@@ -59,3 +59,4 @@
 - while better match result rendering is developed, added sorting to improve readability
     - sorts by match number, then alliance color, to map/render them in sequence of red then blue
     - MUI Accordion (collapsible) per mapped item, to help with visual busyness
+- [Saturday Hotfix] Pit and Match Surveys: change feedback max length to 500 chars